refactor: create packages for mapper and service tests
