+++ domain/src/test/java/com/petdoctor/domain/service/AppointmentServiceTest.java
@@ -1,4 +1,4 @@
-package com.petdoctor.domain;
+package com.petdoctor.domain.service;
 
 import com.petdoctor.data.entity.AppointmentEntity;
 import com.petdoctor.data.entity.AppointmentState;
