+++ domain/src/test/java/com/petdoctor/domain/service/ClientServiceTest.java
@@ -1,4 +1,4 @@
-package com.petdoctor.domain;
+package com.petdoctor.domain.service;
 
 import com.petdoctor.data.entity.ClientEntity;
 import com.petdoctor.data.repository.ClientRepository;
